add funcs to print msg, and add rw ide func (untested)
git-svn-id: https://kenos.googlecode.com/svn/trunk@56 4faa316e-f75d-11dd-a7be-4b3ef3b0700c

+++ include/utils/delay.h
@@ -23,8 +23,8 @@
 #ifndef _KENOS_UT_DELAY_H_
 #define _KENOS_UT_DELAY_H_
 
-static void udelay(unsigned long n);
-static void mdelay(unsigned long n);
+static void udelay(int n);
+static void mdelay(int n);
 
 #endif /* _KENOS_UT_DELAY_H_ */
